Cleaned up dead code. reorganized extensions to all be in one folder.
