+++ StashPacks/StashPacks/Extensions/ContainerSyncPlanExtensions.cs
@@ -1,17 +1,13 @@
-﻿using ModifAmorphic.Outward.StashPacks.SaveData.Extensions;
-using ModifAmorphic.Outward.StashPacks.Sync.Models;
-using System;
-using System.Collections.Generic;
+﻿using ModifAmorphic.Outward.StashPacks.Sync.Models;
 using System.Linq;
-using System.Text;
 
-namespace ModifAmorphic.Outward.StashPacks.Sync.Extensions
+namespace ModifAmorphic.Outward.StashPacks.Extensions
 {
     public static class ContainerSyncPlanExtensions
     {
         public static bool HasChanges(this ContainerSyncPlan syncPlan)
         {
-            return syncPlan.SaveDataAfter.SyncData != syncPlan.SaveDataBefore.SyncData  || syncPlan.AddedItems.Any() || syncPlan.ModifiedItems.Any() || syncPlan.RemovedItems.Any();
+            return syncPlan.SaveDataAfter.SyncData != syncPlan.SaveDataBefore.SyncData || syncPlan.AddedItems.Any() || syncPlan.ModifiedItems.Any() || syncPlan.RemovedItems.Any();
         }
         public static bool HasDifferentSilverAmount(this ContainerSyncPlan syncPlan)
         {
