+++ StashPacks/StashPacks/Extensions/IContainerSaveDataExtensions.cs
@@ -1,11 +1,8 @@
 ﻿using ModifAmorphic.Outward.StashPacks.SaveData.Models;
 using ModifAmorphic.Outward.StashPacks.Sync.Models;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
-namespace ModifAmorphic.Outward.StashPacks.SaveData.Extensions
+namespace ModifAmorphic.Outward.StashPacks.Extensions
 {
     public static class IContainerSaveDataExtensions
     {
@@ -18,7 +15,7 @@ namespace ModifAmorphic.Outward.StashPacks.SaveData.Extensions
                 ContainerType = containerSaveData.ContainerType,
                 ItemID = containerSaveData.ItemID,
                 SaveDataBefore = containerSaveData.BasicSaveData.ToClone(),
-                SaveDataAfter = silver != null ? 
+                SaveDataAfter = silver != null ?
                     containerSaveData.BasicSaveData.ToUpdatedContainerSilver((int)silver) :
                     containerSaveData.BasicSaveData.ToClone(),
                 ItemsSaveDataBefore = containerSaveData.ItemsSaveData.ToDictionary(s => s.Identifier.ToString(), s => s),
