+++ StashPacks/StashPacks/ModifAmorphic.Outward.StashPacks.csproj
@@ -40,10 +40,6 @@
     </Reference>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="SaveData\Sync\" />
-  </ItemGroup>
-
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
     <Exec Command="xcopy $(OutDir)\ModifAmorphic.Outward.StashPacks.dll %25appdata%25\r2modmanPlus-local\Outward\profiles\Default\BepInEx\plugins\ModifAmorphic-ModifAmorphic.Outward.StashPool\ /y" />
   </Target>
