+++ StashPacks/StashPacks/Patch/EnvironmentSavePatches.cs
@@ -1,8 +1,5 @@
 ﻿using HarmonyLib;
 using ModifAmorphic.Outward.StashPacks.Patch.Events;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Patch
 {
