+++ StashPacks/StashPacks/Patch/Events/CharacterInventoryEvents.cs
@@ -1,8 +1,6 @@
 ﻿using ModifAmorphic.Outward.Logging;
-using ModifAmorphic.Outward.StashPacks.WorldInstance.Extensions;
+using ModifAmorphic.Outward.StashPacks.Extensions;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Patch.Events
 {
