+++ StashPacks/StashPacks/Patch/Events/EnvironmentSaveEvents.cs
@@ -1,7 +1,5 @@
 ﻿using ModifAmorphic.Outward.Logging;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Patch.Events
 {
@@ -17,7 +15,7 @@ namespace ModifAmorphic.Outward.StashPacks.Patch.Events
         {
             try
             {
-                Logger?.LogTrace($"{nameof(EnvironmentSaveEvents)}::{nameof(RaiseApplyDataBefore)} raised. EnvironmentSave has {environmentSave?.ItemList?.Count??0} items in it's ItemList.");
+                Logger?.LogTrace($"{nameof(EnvironmentSaveEvents)}::{nameof(RaiseApplyDataBefore)} raised. EnvironmentSave has {environmentSave?.ItemList?.Count ?? 0} items in it's ItemList.");
                 ApplyDataBefore?.Invoke(environmentSave);
             }
             catch (Exception ex)
