+++ StashPacks/StashPacks/Patch/Events/ItemEvents.cs
@@ -1,8 +1,6 @@
 ﻿using ModifAmorphic.Outward.Logging;
 using ModifAmorphic.Outward.StashPacks.Settings;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Patch.Events
 {
