+++ StashPacks/StashPacks/Patch/Events/SaveInstanceEvents.cs
@@ -1,7 +1,5 @@
 ﻿using ModifAmorphic.Outward.Logging;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Patch.Events
 {
