+++ StashPacks/StashPacks/SaveData/Data/StashPackSaveBuilder.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ModifAmorphic.Outward.StashPacks.SaveData.Data
+﻿namespace ModifAmorphic.Outward.StashPacks.SaveData.Data
 {
     class StashPackSaveBuilder
     {
