+++ StashPacks/StashPacks/SaveData/Data/StashPackSaveData.cs
@@ -1,10 +1,5 @@
 ﻿using ModifAmorphic.Outward.Logging;
-using ModifAmorphic.Outward.StashPacks.SaveData.Extensions;
-using ModifAmorphic.Outward.StashPacks.SaveData.Models;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.SaveData.Data
 {
