+++ StashPacks/StashPacks/SaveData/Models/StashPackContainerTypes.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ModifAmorphic.Outward.StashPacks.SaveData.Models
+﻿namespace ModifAmorphic.Outward.StashPacks.SaveData.Models
 {
     public enum ContainerTypes
     {
