+++ StashPacks/StashPacks/SaveData/Models/StashSave.cs
@@ -1,7 +1,4 @@
-﻿using ModifAmorphic.Outward.StashPacks.Sync.Models;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace ModifAmorphic.Outward.StashPacks.SaveData.Models
 {
