+++ StashPacks/StashPacks/Settings/LoggerFactory.cs
@@ -1,7 +1,4 @@
 ﻿using ModifAmorphic.Outward.Logging;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Settings
 {
