+++ StashPacks/StashPacks/Settings/StashPacksConstants.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Settings
 {
