+++ StashPacks/StashPacks/Settings/StashPacksSettings.cs
@@ -1,8 +1,6 @@
 ﻿using ModifAmorphic.Outward.Config.Models;
 using ModifAmorphic.Outward.Logging;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.Settings
 {
