+++ StashPacks/StashPacks/Startup.cs
@@ -4,9 +4,6 @@ using ModifAmorphic.Outward.StashPacks.Patch.Events;
 using ModifAmorphic.Outward.StashPacks.Settings;
 using ModifAmorphic.Outward.StashPacks.State;
 using ModifAmorphic.Outward.StashPacks.WorldInstance.MajorEvents;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks
 {
