+++ StashPacks/StashPacks/State/BagStateService.cs
@@ -1,10 +1,8 @@
 ﻿using ModifAmorphic.Outward.Logging;
-using ModifAmorphic.Outward.StashPacks.SaveData.Extensions;
-using ModifAmorphic.Outward.StashPacks.WorldInstance.Extensions;
+using ModifAmorphic.Outward.StashPacks.Extensions;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Text;
 
 namespace ModifAmorphic.Outward.StashPacks.State
 {
